clean-up upload & download dialogs
